+++ src/types/skinparams.interface.ts
@@ -1,12 +1,12 @@
 import { Direction } from './direction.enum';
 
 export interface SkinParams {
-	direction?: Direction,
-	roundcorner?: number,
-	linetype?: 'ortho' | 'polyline',
-	shadowing?: 'true' | 'false',
-	handwritten?: 'true' | 'false',
 	colors?: Map<string, string>,
+	direction?: Direction,
 	entityNamesOnly?: boolean,
+	handwritten?: 'true' | 'false',
+	linetype?: 'ortho' | 'polyline',
+	roundcorner?: number,
+	shadowing?: 'true' | 'false',
 	tableNamesOnly?: boolean,
 }