+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased] - TBD
 
+- 
+
+## [v1.6.0] (2020-12-10)
+
 - Added `--color` flag to override diagram colors.
 - Added `--plantuml-url` flag to override PlantUML server.
 - Added `--with-entity-names-only` flag to display only entity names in entity titles.
@@ -64,7 +68,8 @@
 - Added `--include` option to include only specific entities to the diagram.
 - Added `--exclude` option to exclude specific entities from the diagram.
 
-[Unreleased]: https://github.com/eugene-manuilov/typeorm-uml/compare/v1.5.0...master
+[Unreleased]: https://github.com/eugene-manuilov/typeorm-uml/compare/v1.6.0...master
+[v1.6.0]: https://github.com/eugene-manuilov/typeorm-uml/compare/v1.5.0...v1.6.0
 [v1.5.0]: https://github.com/eugene-manuilov/typeorm-uml/compare/v1.4.0...v1.5.0
 [v1.4.0]: https://github.com/eugene-manuilov/typeorm-uml/compare/v1.3.0...v1.4.0
 [v1.3.0]: https://github.com/eugene-manuilov/typeorm-uml/compare/v1.2.1...v1.3.0